Minor preferential refactor (#8957)
