+++ pydantic/_internal/_generate_schema.py
@@ -2123,9 +2123,10 @@ def _extract_get_pydantic_json_schema(tp: Any, schema: CoreSchema) -> GetJsonSch
         )
 
         if not has_custom_v2_modify_js_func:
+            cls_name = getattr(tp, '__name__', None)
             raise PydanticUserError(
-                'The `__modify_schema__` method is not supported in Pydantic v2. '
-                'Use `__get_pydantic_json_schema__` instead for class' + str(tp),
+                f'The `__modify_schema__` method is not supported in Pydantic v2. '
+                f'Use `__get_pydantic_json_schema__` instead{f" in class `{cls_name}`" if cls_name else ""}.',
                 code='custom-json-schema',
             )
