+++ tests/test_deprecated.py
@@ -488,13 +488,31 @@ def test_modify_schema_error():
     with pytest.raises(
         PydanticUserError,
         match='The `__modify_schema__` method is not supported in Pydantic v2. '
-        'Use `__get_pydantic_json_schema__` instead.',
+        'Use `__get_pydantic_json_schema__` instead in class `Model`.',
     ):
 
         class Model(BaseModel):
             @classmethod
             def __modify_schema__(cls, field_schema: Dict[str, Any]) -> None:
                 pass
+
+
+def test_modify_schema_on_nested_class_error() -> None:
+    class SomeLongName:
+        @classmethod
+        def __modify_schema__(cls, field_schema):
+            pass
+
+    with pytest.raises(
+        PydanticUserError,
+        match='The `__modify_schema__` method is not supported in Pydantic v2. '
+        'Use `__get_pydantic_json_schema__` instead in class `SomeLongName`.',
+    ):
+
+        class B(BaseModel):
+            model_config = ConfigDict(arbitrary_types_allowed=True)
+
+            a: SomeLongName
 
 
 def test_v1_v2_custom_type_compatibility() -> None:
